Special check overlays for Py2
Will be automatically filled with

    <id>:
        enabled: false
        compatible: false

for Python 2 incompatible plugins on page build, unless `id` is already
present in the data (in case of more specific overlays).

That way plugins will no longer prompt to update to Python 3 only
versions if running on Python 2 (starting with OctoPrint 1.6.0).
Alternatively, plugin authors could rewrite the update URL in such
cases, e.g. to still be notified of security updates from a legacy fork.